+++ air-quality-ui/system_tests/ui/city_page/city_page.spec.ts-snapshots/Midnight-AQI-2-by-PM2-5-chromium-win32.png
