+++ air-quality-ui/system_tests/ui/city_page/city_page.spec.ts-snapshots/Midnight-AQI-3-by-PM2-5-webkit-win32.png
